zsh: separate work-specific configs

+++ bootstrap.sh
@@ -17,16 +17,19 @@ function symlink() {
   mkdir -p "$HOME/bin"
   (cd "$HOME/dotfiles" && stow -v --target="$HOME/bin" bin)
 
-  # VSCode (VSCodium) configs
-  ln -sf "$HOME/dotfiles/.config/vscodium/settings.json"    "$HOME/Library/Application Support/VSCodium/User/settings.json"
-  ln -sf "$HOME/dotfiles/.config/vscodium/keybindings.json" "$HOME/Library/Application Support/VSCodium/User/keybindings.json"
+  # Work-only utilities
+  if [ ! -f "$HOME/.personal_device_marker" ]; then
+    # VSCode (VSCodium) configs
+    ln -sf "$HOME/dotfiles/.config/vscodium/settings.json"    "$HOME/Library/Application Support/VSCodium/User/settings.json"
+    ln -sf "$HOME/dotfiles/.config/vscodium/keybindings.json" "$HOME/Library/Application Support/VSCodium/User/keybindings.json"
 
-  # Kube settings
-  ln -sf "$HOME/dotfiles/.config/kube/kuberc" "$HOME/.kube/kuberc"
+    # Kube settings
+    ln -sf "$HOME/dotfiles/.config/kube/kuberc" "$HOME/.kube/kuberc"
 
-  # OpenCode - symlink to Claude Code's custom commands
-  mkdir -p "$HOME/.config/opencode/command"
-  ln -sf "$HOME/.config/claude/commands" "$HOME/.config/opencode/command/"
+    # OpenCode - symlink to Claude Code's custom commands
+    mkdir -p "$HOME/.config/opencode/command"
+    ln -sf "$HOME/.config/claude/commands" "$HOME/.config/opencode/command/"
+  fi
 }
 
 function dependencies() {
